added multiprocessing and savemodel flag

+++ utils/general_variables.py
@@ -1,4 +1,4 @@
 import torch
 
-DATA_TYPE = torch.cuda.DoubleTensor
-# DATA_TYPE = torch.cuda.FloatTensor
+# DATA_TYPE = torch.cuda.DoubleTensor
+DATA_TYPE = torch.cuda.FloatTensor
